Decouple bound endpoint allocator and endpoint container.
This moves the bound allocator into Connection, and removes an unused method.
